+++ NWNXLib/API/API/CNWSArea.hpp
@@ -24,6 +24,7 @@ struct CERFFile;
 struct CGameEffect;
 struct CGameObject;
 struct CNWSAmbientSound;
+struct CNWSAreaGridPoint;
 struct CNWSAreaGridSuccessors;
 struct CNWSAreaGridTransTableEntry;
 struct CNWSAreaInterTileSuccessors;
@@ -139,6 +140,9 @@ struct CNWSArea : CNWArea, CResHelper<CResARE, 2012>, CGameObject
     virtual BOOL NoCreaturesOnLine(float fSourceX, float fSourceY, float fTargetX, float fTargetY, CPathfindInformation * pPathfindInfo, BOOL bCheckSourceOccluded = true, BOOL bIgnoreAssociates = false, OBJECT_ID * poidBlockingCreature = nullptr, BOOL bEvaluateOverlappingTarget = false);
     BOOL EvaluateOverlappingTargets(CPathfindInformation * pPathfindInfo, Vector vPosition, OBJECT_ID oidAreaCreature, float fSourceX, float fSourceY, float fBothCreaturesPersonalSpace, BOOL bCheckSourceOccluded, BOOL bEvaluateOverlappingTarget, OBJECT_ID * poidBlockingCreature);
     BOOL PackAreaIntoMessage(int32_t nX, int32_t nY, int32_t nZ, CNWSPlayer * pPlayer);
+    uint32_t PlotGridPathEnhanced(CPathfindInformation * pcPathfindInformation, uint64_t nTimeSlice, bool bFinalAttempt);
+    void GridSearchPath(int32_t nPathID);
+    bool GridSearchPathInDirection(int32_t nStepNumber, int32_t nX, int32_t nY, int nDirectionX, int nDirectionY, bool bTestDirectLine, CNWSAreaGridPoint * pcSearch);
     uint32_t PlotGridPath(CPathfindInformation * pcPathfindInformation, uint64_t nTimeSlice);
     BOOL GridDFSearch(int32_t nStepNumber, int32_t nPathLength, int32_t nStepsToExplore, int32_t nX, int32_t nY, int32_t nZ);
     int32_t GridDFSGenerateSuccessors(int32_t nPathLength, int32_t nStepsToExplore, int32_t nX, int32_t nY, int32_t nZ, CNWSAreaGridSuccessors * * pcSuccessors);